+++ packages/@eventual/integrations-slack/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.2.0 (2022-12-21)
+
+### Features
+
+- support eventual-infer on .ts using esbuild ([#132](https://github.com/functionless/eventual/issues/132)) ([0f4259f](https://github.com/functionless/eventual/commit/0f4259f0ca42efcd5ff2803af11bee2a46ed0186))
+
+## 0.1.1 (2022-12-21)
+
+### Bug Fixes
+
+- take peer on ^2.50.0 or AWS CDK ([#126](https://github.com/functionless/eventual/issues/126)) ([fe850bb](https://github.com/functionless/eventual/commit/fe850bb180bb7e599fae91346b244cd0e50cbbe3))
+
+# 0.1.0 (2022-12-21)
+
+### Features
+
+- add signal function ([#129](https://github.com/functionless/eventual/issues/129)) ([8c43568](https://github.com/functionless/eventual/commit/8c43568a19c36a2792d440520d9606d03c57e606))
+
 ## 0.0.1 (2022-12-19)
 
 **Note:** Version bump only for package @eventual/integrations-slack